Consolidate type data into `types.ts`
