+++ packages/components/src/color-picker/legacy-adapter.tsx
@@ -2,9 +2,8 @@
  * Internal dependencies
  */
 import ColorPicker from './component';
+import type { LegacyAdapterProps } from './types';
 import { useDeprecatedProps } from './use-deprecated-props';
-
-type LegacyAdapterProps = Parameters< typeof useDeprecatedProps >[ 0 ];
 
 export const LegacyAdapter = ( props: LegacyAdapterProps ) => {
 	return <ColorPicker { ...useDeprecatedProps( props ) } />;
